offset and size completed(union,struct,array cases handled).

+++ tests/test6.c
@@ -21,7 +21,7 @@ int main(void)
 {
     int arr[] = { 2, 3, 4, 10, 40 };
     int n = sizeof(arr) / sizeof(arr[0]);
-    int x = 10;
+    double x = 10;
     int result = binarySearch(arr, 0, n - 1, x,1);
     return 0;
 }
